Added support for All Domains domains
