+++ app/components/account/DomainsCard.tsx
@@ -4,22 +4,36 @@ import { Address } from '@components/common/Address';
 import { ErrorCard } from '@components/common/ErrorCard';
 import { LoadingCard } from '@components/common/LoadingCard';
 import { useUserDomains } from '@utils/name-service';
+import { useUserANSDomains } from "../../utils/ans-domains";
+
 import React from 'react';
 
 import { DomainInfo } from '@/app/utils/domain-info';
 
 export function DomainsCard({ address }: { address: string }) {
     const [domains, domainsLoading] = useUserDomains(address);
+    const [domainsANS, domainsANSLoading] = useUserANSDomains(address);
 
-    if (domainsLoading && (!domains || domains.length === 0)) {
+    if (
+      (domainsLoading && (!domains || domains.length === 0)) ||
+      (domainsANSLoading && (!domainsANS || domainsANS.length === 0))
+    ) {
         return <LoadingCard message="Loading domains" />;
-    } else if (!domains) {
+    } else if (!domains || !domainsANS) {
         return <ErrorCard text="Failed to fetch domains" />;
     }
 
-    if (domains.length === 0) {
+    if (domains.length === 0 && domainsANS.length === 0) {
         return <ErrorCard text="No domain name found" />;
     }
+
+    let allDomains = domains;
+
+    if (domainsANS) {
+      allDomains = [...allDomains, ...domainsANS];
+    }
+
+    allDomains.sort((a, b) => a.name.localeCompare(b.name));
 
     return (
         <div className="card">
@@ -35,7 +49,7 @@ export function DomainsCard({ address }: { address: string }) {
                         </tr>
                     </thead>
                     <tbody className="list">
-                        {domains.map(domain => (
+                        {allDomains.map((domain) => (
                             <RenderDomainRow key={domain.address.toBase58()} domainInfo={domain} />
                         ))}
                     </tbody>
